Move JetStreamOptions into jetstream module
